[lsm] add Catch2 based test system
I also added some flair to our "integration tests" (test.sh).
