+++ docker/Dockerfile.base
@@ -13,7 +13,8 @@ RUN hail-apt-get-install \
     openjdk-8-jdk-headless \
     liblapack3 \
     g++-10 \
-    gcc-10
+    gcc-10 \
+    catch
 
 RUN update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-10 10 && \
     update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-10 10
