+++ lsm/README.md
@@ -1,0 +1,4 @@
+# Compile-Time Dependencies
+
+- [Catch2](https://github.com/catchorg/Catch2/blob/devel/docs/tutorial.md#top). It
+  is available in Homebrew as `catch2`.
